backend/requirements.txt
aioboto3==15.4.0
aiobotocore==2.25.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aioitertools==0.12.0
aiosignal==1.4.0
amazon-bedrock-haystack==5.2.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
boto3==1.40.49
botocore==1.40.49
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
duckdb==1.4.1
environs==14.3.0
fastapi==0.120.0
fastapi_cors==0.0.6
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.42.0
google-generativeai
google-genai-haystack==2.2.0
h11==0.16.0
haystack-ai==2.19.0
haystack-experimental==0.14.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_imports==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
openai==2.6.1
packaging==25.0
pillow==12.0.0
posthog==6.7.11
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyMuPDF==1.26.6
pypdf==6.1.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
yarl==1.22.0
pywin32==308
